I think I know how to build the array :D
I think I know how to build the array :D Take a look at this. That is how to sum the third column. I did not resolve the problem how to build this new array based on user picks, but at least I know how to make it work to sum particular column values. Not too much, but always a progress.